[llvm, Darwin, PPC] Allow 'ppc' internally, as an arch name.
This is a compromise; although the arch is ppc, the user-visible triple
is called powerpc-apple-darwin.  However, within llvm the arch portion
of the triple is often compared with the user's value for '-arch' (where
the correct spelling is 'ppc').  I guess we should invent a 'normalise
triple for presentation to the user' function.